core: hook: add functions to convert nf hooks back and forth
Two functions are used to convert nf hooks to bf hook and reverse. Move
those functions to hook.h where they belong.
