+++ .clang-tidy
@@ -9,6 +9,8 @@
 # - bugprone-easily-swappable-parameters
 #   Too many false positives, especially when swapable arguments are of different
 #   types (which will be flagged by the compiler).
+# - cert-dcl03-c
+#   We use bf_assert(0) for impossible default in switch...case.
 # - cert-dcl37-c
 #   Handled by bugprone-reserved-identifier.AllowedIdentifiers.
 # - cert-dcl51-cpp
@@ -31,6 +33,8 @@
 #   Avoid usage of Annex K functions for portability reasons.
 # - clang-analyzer-unix.Malloc
 #   Generates false positives.
+# - misc-static-assert
+#   We use bf_assert(0) for impossible default in switch...case.
 # - modernize-macro-to-enum
 #   No benefit.
 # - modernize-use-trailing-return-type
@@ -54,6 +58,7 @@ Checks: >
     -bugprone-assignment-in-if-condition,
     -bugprone-easily-swappable-parameters,
   cert-*,
+    -cert-dcl03-c,
     -cert-dcl37-c,
     -cert-dcl51-cpp,
     -cert-int09-c,
@@ -66,6 +71,7 @@ Checks: >
     -clang-analyzer-security.insecureAPI.DeprecatedOrUnsafeBufferHandling,
     -clang-analyzer-unix.Malloc,
   misc-*,
+    -misc-static-assert,
   modernize-*,
     -modernize-macro-to-enum,
     -modernize-use-trailing-return-type,
